Fix registration
- Moved create, store and validation methods from EmployeeController to RegistrationController.
- Dots on the i within RegistrationController.
- Adjusted routes to RegistrationController.
- Removed employee resource route and added them within the employee middleware.
